Squashed commit of the following:
commit 55f159e6034dc874bcd5fc410b485567ec0c9c53
Author: Khang Duong <khangvngm@gmail.com>
Date:   Tue Nov 21 22:18:00 2023 +0700

    Update 21/11/2023

commit a2445325f14ce4f7734101ff7a8ee8f95dd2e376
Author: Khang Duong <khangvngm@gmail.com>
Date:   Mon Nov 20 23:08:20 2023 +0700

    no message

commit 7daafffce81952a893a8f31ce18378cff9c5f18c
Author: Khang Duong <khangvngm@gmail.com>
Date:   Mon Nov 20 21:49:08 2023 +0700

    no message

commit 3560d68f026cff3432913f930aba585114c42a11
Author: Khang Duong <khangvngm@gmail.com>
Date:   Fri Nov 17 22:38:20 2023 +0700

    Update 11/17/2023

commit c488d50ce01eca085980925075fcaa41ab8e882e
Author: Khang Duong <khangvngm@gmail.com>
Date:   Tue Nov 14 21:39:51 2023 +0700

    no message

commit 2f045adc90df970330fc10d4558cda41df1c9ebc
Author: Khang Duong <khangvngm@gmail.com>
Date:   Sat Nov 11 23:41:51 2023 +0700

    no message

commit 36f4b87e7e95bda539df9b7fc191c005983ae9ab
Author: Khang Duong <khangvngm@gmail.com>
Date:   Sat Nov 11 22:28:23 2023 +0700

    Update extra jump, jump in wall, attack up - down and  holy light skill
    Fix forward vs restreat attack,

commit 4dd701176a6cfdb4cf14973daa0ebac35a7b343a
Merge: 653261e d12832c
Author: Khang Duong <khangvngm@gmail.com>
Date:   Tue Nov 7 19:01:37 2023 +0700

    Update kamehameha vs fireball

commit 653261e896618417d72d58d5f26e8f86703e8ad4
Author: Khang Duong <khangvngm@gmail.com>
Date:   Tue Nov 7 00:49:07 2023 +0700

    no message

commit d12832c2ff9de35f5e10c1da10f2f877caefeb57
Author: hvqDevUnity <hoangvanquan2002.unitydevelop@gmail.com>
Date:   Mon Nov 6 23:27:57 2023 +0700

    fix bug load scene

commit 88cad1c26c2e9b446ce16fdb5c456417bb515424
Author: hvqDevUnity <hoangvanquan2002.unitydevelop@gmail.com>
Date:   Mon Nov 6 23:23:30 2023 +0700

    bug load submit

commit 58e00284b449de2ba649adbdd82778e7401e7bf5
Author: Khang Duong <khangvngm@gmail.com>
Date:   Mon Nov 6 23:09:13 2023 +0700

    no message

commit 35fd3d388eee6f849255067de3262b2bffc29904
Author: hvqDevUnity <hoangvanquan2002.unitydevelop@gmail.com>
Date:   Mon Nov 6 23:08:20 2023 +0700

    store finish

commit e08ec65e0b5bd1fca58567d822fc52a38f56c87e
Author: hvqDevUnity <hoangvanquan2002.unitydevelop@gmail.com>
Date:   Mon Nov 6 22:52:01 2023 +0700

    store

commit 2afb49b23b119707791b45c74b6d7165dd6c92db
Author: Khang Duong <khangvngm@gmail.com>
Date:   Mon Nov 6 21:56:07 2023 +0700

    Update Fix Concentrate Skill

commit 7193388d672fcdf6aaea7014b5758124010886e1
Author: hvqDevUnity <hoangvanquan2002.unitydevelop@gmail.com>
Date:   Mon Nov 6 21:38:43 2023 +0700

    add tags to support skill

commit dfe47d11e0318c47657f9685e2c05473ec0bcfc7
Author: Khang Duong <khangvngm@gmail.com>
Date:   Mon Nov 6 19:01:45 2023 +0700

    Update Nothingness & Contrentrate Skill

commit d813bd2d1ec059670031adf5bf5cea658d9ec599
Author: Khang Duong <khangvngm@gmail.com>
Date:   Mon Nov 6 01:10:58 2023 +0700

    no message

commit 81ba50f119ba4ba2390856b1d766bdf8af59b0cc
Merge: 0d8d065 d0f4d4a
Author: Khang Duong <khangvngm@gmail.com>
Date:   Mon Nov 6 00:47:04 2023 +0700

    Merge branch 'local02' of https://github.com/shiroeyamamoto/Evil-Collectors into local

    # Conflicts:
    #	Assets/Prefabs/Player/Skill/Nothingness.prefab
    #	Assets/Prefabs/Player/Skill/Nothingness.prefab.meta
    #	ProjectSettings/ProjectVersion.txt
    #	UserSettings/EditorUserSettings.asset
    #	UserSettings/Layouts/CurrentMaximizeLayout.dwlt
    #	UserSettings/Layouts/default-2022.dwlt
    #	UserSettings/Search.settings

commit d0f4d4aa8d9620ac85d777e946e102dcbbf52c07
Merge: 4d6bad2 44acd62
Author: Khang Duong <khangvngm@gmail.com>
Date:   Mon Nov 6 00:39:51 2023 +0700

    Merge commit '44acd62249142138b817d123768812f99b65e37e' into local02

    # Conflicts:
    #	Assets/Scenes/PlayerTest.unity
    #	Assets/Scripts/Player/Player.cs

commit 0d8d06536fa3a08aadf4df94153d735015de85b9
Author: Khang Duong <khangvngm@gmail.com>
Date:   Mon Nov 6 00:24:07 2023 +0700

    no message

commit 4d6bad279bb355ea54f62e8de84b2ad2c805f299
Merge: 09cbdb0 0af0d8b
Author: hvqDevUnity <hoangvanquan2002.unitydevelop@gmail.com>
Date:   Mon Nov 6 00:21:47 2023 +0700

    Merge remote-tracking branch 'origin/main' into develop_Quan

    # Conflicts:
    #	Assets/Scripts/Interface/IInteractObject.cs
    #	Assets/Scripts/Player/Player.cs

commit 09cbdb0f2a1a8cfd8d05c58a0a5833bbaf7998a8
Author: hvqDevUnity <hoangvanquan2002.unitydevelop@gmail.com>
Date:   Sun Nov 5 23:42:20 2023 +0700

    merge

commit 44acd62249142138b817d123768812f99b65e37e
Author: Khang Duong <khangvngm@gmail.com>
Date:   Sun Nov 5 23:27:50 2023 +0700

    no message

commit 9a38ac5617c21fe3fba0ef019561f037e9253900
Merge: f1fb718 85ce8f9
Author: hvqDevUnity <hoangvanquan2002.unitydevelop@gmail.com>
Date:   Sun Nov 5 22:52:12 2023 +0700

    Merge remote-tracking branch 'origin/main' into develop_Quan

    # Conflicts:
    #	Assets/NguyenFolder/Script/BossController.cs
    #	Assets/Scripts/Enemys/EnemyAttack.cs
    #	Assets/Scripts/Player/Player.cs
    #	Assets/Scripts/Player/PlayerCollision.cs

commit f1fb7186644390ac9d0748e93042a7a3d1812152
Merge: d4bae29 9606232
Author: hvqDevUnity <hoangvanquan2002.unitydevelop@gmail.com>
Date:   Sun Nov 5 20:57:11 2023 +0700

    Merge branch 'dev' into develop_Quan

    # Conflicts:
    #	Assets/Scenes/MenuScene.unity
    #	Assets/Scenes/PersistentScene.unity

commit 96062323a76efe4180743e1c21d1f3a863eb1276
Author: hvqDevUnity <hoangvanquan2002.unitydevelop@gmail.com>
Date:   Sun Nov 5 20:47:13 2023 +0700

    missing prefabs

commit d4bae29234f653367a85b35b0fbab70fe7e2d54c
Merge: 7b6846d 51f49a9
Author: hvqDevUnity <hoangvanquan2002.unitydevelop@gmail.com>
Date:   Sun Nov 5 20:44:13 2023 +0700

    Merge remote-tracking branch 'origin/main' into develop_Quan

    # Conflicts:
    #	Assets/Scenes/MenuScene.unity
    #	Assets/Scenes/PersistentScene.unity
    #	Assets/Scripts/Player/Player.cs
    #	UserSettings/Layouts/default-2022.dwlt

commit 7b6846df8ecc66e48f69eff739a6f255c8646532
Author: hvqDevUnity <hoangvanquan2002.unitydevelop@gmail.com>
Date:   Sun Nov 5 20:35:30 2023 +0700

    Skill, Load Skill, Buff Skill from Support Item

commit 40e628b8d3dc68420e13417a38c7e4d63f2ef6e2
Merge: b489e7d cac6115
Author: hvqDevUnity <hoangvanquan2002.unitydevelop@gmail.com>
Date:   Sun Nov 5 18:53:58 2023 +0700

    Merge branch 'dev' into develop_Quan

    # Conflicts:
    #	Assets/Scenes/MenuScene.unity
    #	Assets/ScriptableObject/Items/a1_CauLua.asset.meta
    #	Assets/ScriptableObject/Items/a2_Chuong.asset
    #	Assets/ScriptableObject/Items/i4_poisionKnight.asset
    #	Assets/ScriptableObject/Items/i4_poisionKnight.asset.meta
    #	Assets/ScriptableObject/Items/i4_posionKnight.asset
    #	Assets/ScriptableObject/Items/i4_posionKnight.asset.meta
    #	Assets/Scripts/Data/ItemSO.cs
    #	UserSettings/Layouts/default-2022.dwlt

commit b489e7d922658e6c55bd4926cb01c8f6251cb1b1
Author: hvqDevUnity <hoangvanquan2002.unitydevelop@gmail.com>
Date:   Sun Nov 5 18:51:13 2023 +0700

    item base, support item, active item, Consumable Item, skill base , kamehameha, firesphere

commit cac61155793daace1326006acc4c9a639de9f121
Author: DESKTOP-7Q0HBGC\HLC_2021 <hoangvanquan2002.unitydevelop@gmail.com>
Date:   Sat Nov 4 18:25:16 2023 +0700

    Circle Logic UI

commit 4c69cc5fb672d5ea95380dbfa08d147a79dea444
Author: hvqDevUnity <hoangvanquan2002.unitydevelop@gmail.com>
Date:   Sat Nov 4 13:58:56 2023 +0700

    item SO

+++ Assets/Scripts/Data/ConsumableItem.cs
@@ -41,7 +41,7 @@ public class ConsumableItem : ItemBase
     private void DoUse(Player player) {
         if (player.CurrentInfo.health > 0) {
             if (restoreMaxHP) {
-                player.IncreaseHp((int)player.InfoDefaultSO.health);
+                player.IncreaseHp(player.InfoDefaultSO.health);
             }
             else {
                 player.IncreaseHp(hpAmount);
@@ -62,7 +62,7 @@ public class ConsumableItem : ItemBase
         {
             if (restoreMaxHP)
             {
-                Player.Instance.IncreaseHp((int)Player.Instance.InfoDefaultSO.health);
+                Player.Instance.IncreaseHp(Player.Instance.InfoDefaultSO.health);
             }
             else
             {
